Have a partial model for the engine.
I am using multipile enumerated types to store units.  At the moment, it
is only length.  There will be a field for unit type when I allow for
selection between length, area, volume, and temperal.

Wrote some basic unit tests for making the conversion.  They don't all
work yet, but I want to get a basic start.

The idea is, you walk your way up or down the list of enums, multiplying
or dividing as you go.  I will need to add a field for the value behind.
Also, the conversion needs to handle the not a unit case.