Took out unneeded Annotation.  Added unit tests.
Can select a property an display it in text.

Next, want to have a list in different cells.

We want test coverage for the basic running of conversion and the
converting code.  This shows that the proper data gets put into the
output map.  Also shows that the length units do get put in the map.